Fix masterdata formatting
