+++ src/FasTnT.Data.PostgreSql/DTOs/Masterdata/MasterDataHierarchyDto.cs
@@ -17,5 +17,10 @@ namespace FasTnT.Data.PostgreSql.DTOs
                 ChildrenId = children
             };
         }
+
+        internal string ToHierarchy()
+        {
+            return ChildrenId;
+        }
     }
 }
