+++ src/FasTnT.Data.PostgreSql/DapperConfiguration/SqlQueries.resx
@@ -163,6 +163,9 @@ DROP TABLE eventIds;</value>
   <data name="Read_MasterdataChildren" xml:space="preserve">
     <value>SELECT type, parent_id, children_id FROM cbv.hierarchy WHERE parent_id = ANY(@Ids);</value>
   </data>
+  <data name="Read_UserByUsername" xml:space="preserve">
+    <value>SELECT id, username, password FROM users.user WHERE username = @username</value>
+  </data>
   <data name="Store_ContactInformationDto" xml:space="preserve">
     <value>INSERT INTO sbdh.contactinformation(id, header_id, type, identifier, contact, email, fax_number, phone_number, type_identifier) VALUES(@id, @headerid, @type, @identifier, @contact, @emailaddress, @faxnumber, @telephonenumber, @contacttypeidentifier) ...;</value>
   </data>
