--- src/FasTnT.Domain/Data/Model/CaptureDocumentRequest.cs
@@ -1,9 +1,0 @@
-﻿using FasTnT.Model;
-
-namespace FasTnT.Domain.Data.Model
-{
-    public class CaptureDocumentRequest
-    {
-        public EpcisRequest Header { get; set; }
-    }
-}
