--- src/FasTnT.Domain/Model/Events/ErrorDeclaration.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace FasTnT.Model.Events
-{
-    public class ErrorDeclaration
-    {
-        public DateTime? DeclarationTime { get; set; }
-        public string Reason { get; set; }
-    }
-}
